# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulMatch AI - AI-powered matrimony platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send email/phone OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/google:
    post:
      operationId: googleLogin
      tags: [auth]
      summary: Google OAuth login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /chat/conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chat/conversations/{conversationId}/read:
    post:
      operationId: markConversationRead
      tags: [chat]
      summary: Mark all messages in conversation as read
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, dateOfBirth, gender]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SendOtpInput:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [email, phone]
        value:
          type: string
    VerifyOtpInput:
      type: object
      required: [type, value, otp]
      properties:
        type:
          type: string
          enum: [email, phone]
        value:
          type: string
        otp:
          type: string
    GoogleLoginInput:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── USER SCHEMAS ────────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        motherTongue:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        height:
          type: ["integer", "null"]
        maritalStatus:
          type: ["string", "null"]
        dietaryPreference:
          type: ["string", "null"]
        smoking:
          type: ["string", "null"]
        drinking:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        role:
          type: string
          enum: [user, premium, admin, superadmin]
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isPhoneVerified:
          type: boolean
        isEmailVerified:
          type: boolean
        journeyProgress:
          type: integer
        isPremium:
          type: boolean
        profileCompleteness:
          type: integer
        createdAt:
          type: string
    Photo:
      type: object
      required: [id, url, isPrimary]
      properties:
        id:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
        publicId:
          type: ["string", "null"]
...
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        messageType:
          type: string
          enum: [text, image, emoji]
        isRead:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        messageType:
          type: string
          enum: [text, image, emoji]

    # ── SUBSCRIPTION SCHEMAS ────────────────────────────────────────────────
...